requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blis==1.3.0
Brotli==1.1.0
bs4==0.0.2
cairocffi==1.7.1
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
diff-match-patch==20241021
diffusers==0.33.1
distro==1.9.0
Django==4.2.19
django-admin-rangefilter==0.13.2
django-allauth==65.8.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-extensions==4.1
django-import-export==4.3.7
django-js-asset==3.1.2
django-maintenance-mode==0.22.0
djangorestframework==3.16.0
drf-yasg==1.21.10
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonfield==3.1.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
murmurhash==1.0.13
mysqlclient==2.2.7
narwhals==1.40.0
networkx==3.4.2
numpy==2.2.6
openai==0.28.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdfkit==1.0.0
pillow==11.2.1
playwright==1.52.0
plotly==6.1.1
ply==3.11
preshed==3.0.10
propcache==0.3.1
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMuPDF==1.26.0
pyparsing==3.2.3
pypdf==5.6.0
pyphen==0.17.2
pysmi==1.6.1
pysnmp==4.4.12
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.1.0
python-fsutil==0.15.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
razorpay==1.4.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scapy==2.6.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.33.0
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speedtest-cli==2.1.3
sqlparse==0.5.3
srsly==2.5.1
svglib==1.5.1
sympy==1.14.0
tablib==3.8.0
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.2
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
uritools==5.0.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==53.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xhtml2pdf==0.2.17
yarl==1.20.0
zipp==3.22.0
zopfli==0.2.3.post1
